Create README.md
The Project is about to find the spammer groups who are working for the same products to promote or demote it. I have applied different Machine learning, deep learning and Natural language processing techniques to detect the groups of spam people. The two datasets are used yelp and amazon. The demonstration of ML model is done through django and model pickle file is saved in the django through which predicted result is obtained and we can see our result in the webpage.